+++ app/scripts/fhir/R4/lformsFHIR.js
@@ -19153,10 +19153,11 @@ var sdcExport = {
           "valueBoolean": true
         });
       }
-    } // calcuatedValue
+    } // Copied item extensions
 
 
-    if (item._calculatedExprExt) targetItem.extension.push(item._calculatedExprExt); // http://hl7.org/fhir/StructureDefinition/questionnaire-itemControl
+    if (item._calculatedExprExt) targetItem.extension.push(item._calculatedExprExt);
+    if (item._initialExprExt) targetItem.extension.push(item._initialExprExt); // http://hl7.org/fhir/StructureDefinition/questionnaire-itemControl
 
     this._handleItemControl(targetItem, item); // questionnaire-choiceOrientation , not supported yet
     // check restrictions
@@ -20557,7 +20558,7 @@ function addSDCImportFns(ns) {
 
     self._processSkipLogic(targetItem, qItem, qResource);
 
-    self._processCalculatedValue(targetItem, qItem);
+    self._processCopiedItemExtensions(targetItem, qItem);
 
     self.copyFields(qItem, targetItem, self.itemLevelIgnoredFields);
 
@@ -20572,18 +20573,28 @@ function addSDCImportFns(ns) {
     }
 
     return targetItem;
+  }; // A map of FHIR extensions involving Expressions to the property names on
+  // which they will be stored in LFormsData.
+
+
+  var expressionExtensions = {
+    "http://hl7.org/fhir/StructureDefinition/questionnaire-calculatedExpression": "_calculatedExprExt",
+    "http://hl7.org/fhir/StructureDefinition/questionnaire-initialExpression": "_initialExprExt"
   };
+  var expressionExtURLs = Object.keys(expressionExtensions);
   /**
-   *  Copies the calculated value expression from qItem to lfItem if it exists,
-   *  and if it is a FHIRPath expression, which is the only type we support.
+   *  Some extensions are simply copied over to the LForms data structure.
+   *  This copies those extensions from qItem to lfItem if they exist, and
+   *  LForms can support them.
+   * @param qItem an item from the Questionnaire resource
+   * @param lfItem an item from the LFormsData structure
    */
 
-
-  self._processCalculatedValue = function (lfItem, qItem) {
-    var calcExt = LForms.Util.findObjectInArray(qItem.extension, 'url', "http://hl7.org/fhir/StructureDefinition/questionnaire-calculatedExpression");
-
-    if (calcExt && calcExt.valueExpression.language == "text/fhirpath") {
-      lfItem._calculatedExprExt = calcExt;
+  self._processCopiedItemExtensions = function (lfItem, qItem) {
+    for (var i = 0, len = expressionExtURLs.length; i < len; ++i) {
+      var url = expressionExtURLs[i];
+      var ext = LForms.Util.findObjectInArray(qItem.extension, 'url', url);
+      if (ext && ext.valueExpression && ext.valueExpression.language === "text/fhirpath") lfItem[expressionExtensions[url]] = ext;
     }
   };
   /**