+++ test/karma/fhir-sdc.spec.js
@@ -281,6 +281,48 @@ for (var i=0, len=fhirVersions.length; i<len; ++i) {
               fhirQ.item[0].extension[0].url);
           });
 
+          if(fhirVersion === 'STU3') {
+            describe.only('argonaut samples', function () {
+              it('should parse housing', function (done) {
+                var file = 'test/data/STU3/argonaut-examples/housing.json';
+                $.get(file, function(json) {
+                  var lfData = LForms.Util.convertFHIRQuestionnaireToLForms(json, fhirVersion);
+                  var convertedQ = LForms.Util._convertLFormsToFHIRData('Questionnaire', fhirVersion, lfData);
+                  assert.equal(convertedQ.item[0].item[1].option.length, json.item[0].item[1].option.length);
+                  assert.equal(convertedQ.item[0].item[2].option.length, json.item[0].item[2].option.length);
+
+                  // valueString is changed to valueCoding.display
+                  assert.equal(convertedQ.item[0].item[1].option[0].valueCoding.display, json.item[0].item[1].option[0].valueString);
+
+                }).done(function () {
+                  done();
+                }).fail(function (err) {
+                  done(err);
+                });
+              });
+
+              it.only('should parse sampler', function (done) {
+                var file = 'test/data/STU3/argonaut-examples/sampler.json';
+                $.get(file, function(json) {
+                  var lfData = LForms.Util.convertFHIRQuestionnaireToLForms(json, fhirVersion);
+                  var convertedQ = LForms.Util._convertLFormsToFHIRData('Questionnaire', fhirVersion, lfData);
+
+                  assert.equal(convertedQ.item[11].item[0].option.length, json.item[11].item[0].option.length);
+                  // The score is changed from argonaut extension to FHIR extension.
+                  assert.equal(convertedQ.item[11].item[0].option[0].extension[0].url,
+                    'http://hl7.org/fhir/StructureDefinition/questionnaire-ordinalValue');
+                  assert.equal(convertedQ.item[11].item[0].option[0].extension[0].valueDecimal,
+                    json.item[11].item[0].option[0].extension[0].valueDecimal);
+                }).done(function () {
+                  done();
+                }).fail(function (err) {
+                  done(err);
+                });
+              });
+            });
+          }
+
+
         });
 
         describe('LForms data to QuestionnaireResponse conversion', function() {