+++ CHANGELOG.md
@@ -2,9 +2,18 @@
 
 This project follows [Semantic Versioning](http://semver.org/).
 
-## [13.6.0] 2018-12-03
+## [13.7.1] 2018-12-07
+### Fixed.
+- Fixed errors in parsing argonaut questionnaire extensions. 
+
+## [13.7.0] 2018-12-06
 ### Added.
 - Added support for enableWhen and initial(x) in R4. 
+
+## [13.6.0] 2018-12-04
+### Added
+- Support the FHIR extension "questionnaire-initialExpression".
+
 ## [13.5.0] 2018-11-26
 ### Added
 - A new function LForms.Util.guessFHIRVersion(fhirResource) which looks at the